% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{latent_heat_vap}
\alias{latent_heat_vap}
\title{latent heat of vaporization (lambda) at normal pressure J/kg}
\usage{
latent_heat_vap()
}
\description{
latent heat of vaporization (lambda) at normal pressure J/kg
}
