% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{O}
\alias{O}
\title{potential temperature,  k=R/cp, after Arya, S. P. (2001). Introduction to Micrometeorology. International Geophysics Series (Vol. 79, p. 420). San Diego: Academic Press. doi:10.4043/13298-MS))}
\usage{
O(T, p)
}
\description{
potential temperature,  k=R/cp, after Arya, S. P. (2001). Introduction to Micrometeorology. International Geophysics Series (Vol. 79, p. 420). San Diego: Academic Press. doi:10.4043/13298-MS))
}
