% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{psychr}
\alias{psychr}
\title{psychrometric constant in kPa/K for local pressure conditions
check http://www.fao.org/docrep/X0490E/x0490e07.htm#psychrometric%20constant%20(g)}
\usage{
psychr(p)
}
\description{
psychrometric constant in kPa/K for local pressure conditions
check http://www.fao.org/docrep/X0490E/x0490e07.htm#psychrometric%20constant%20(g)
}
