% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_COMBILOG.R
\name{read_COMBILOG}
\alias{read_COMBILOG}
\title{read logger file from theodor friedrichs}
\usage{
read_COMBILOG(path)
}
\arguments{
\item{path}{is the absolute path to the logger file}
}
\description{
read logger file from theodor friedrichs
}
