% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{rh2specific_hum}
\alias{rh2specific_hum}
\title{relative humidity to specific humidity}
\usage{
rh2specific_hum(rh, T, p)
}
\arguments{
\item{rh}{in [-]}

\item{T}{in [K]}

\item{p}{in kPa}
}
\description{
relative humidity to specific humidity
}
