% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{rh2vpressure}
\alias{rh2vpressure}
\title{relative humidity to vapour pressure}
\usage{
rh2vpressure(rh, T)
}
\arguments{
\item{rh}{in [-]}

\item{T}{in [K]}
}
\description{
relative humidity to vapour pressure
}
