% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{sat_vpressure}
\alias{sat_vpressure}
\title{vapour pressure at saturation in kPa after Moene and Van Dam Cambridge University Press 2014 "Transport in the Atmosphere-Vegetation-Soil Continuum " page 353 eq B.20}
\usage{
sat_vpressure(T)
}
\arguments{
\item{T}{in [K]}

\item{esat}{in [kPa]}
}
\description{
vapour pressure at saturation in kPa after Moene and Van Dam Cambridge University Press 2014 "Transport in the Atmosphere-Vegetation-Soil Continuum " page 353 eq B.20
}
