% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{slope_sat_vpressure}
\alias{slope_sat_vpressure}
\title{slope of the saturation vapour pressure-temperature relationship, from Allen et al, FAO crop water requirements (1998)}
\usage{
slope_sat_vpressure(T)
}
\arguments{
\item{T}{in K}
}
\description{
slope of the saturation vapour pressure-temperature relationship, from Allen et al, FAO crop water requirements (1998)
}
