% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{spec_heat_cp_air}
\alias{spec_heat_cp_air}
\title{specific heat capacity of air at constant pressure}
\usage{
spec_heat_cp_air(q)
}
\arguments{
\item{q}{is specific humidity [kg.kg-1]}
}
\description{
specific heat capacity of air at constant pressure
}
