% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{specific_hum2rh}
\alias{specific_hum2rh}
\title{relative humidity after Moene and Van Dam Cambridge University Press 2014 "Transport in the Atmosphere-Vegetation-Soil Continuum " page 352}
\usage{
specific_hum2rh(q, T, p)
}
\arguments{
\item{q}{in [kg.kg^-1]}

\item{T}{in [K]}

\item{p}{in Pa}
}
\description{
relative humidity after Moene and Van Dam Cambridge University Press 2014 "Transport in the Atmosphere-Vegetation-Soil Continuum " page 352
}
