% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{specific_hum2vpressure}
\alias{specific_hum2vpressure}
\title{vapour pressure in Pa after Moene and Van Dam Cambridge University Press 2014 "Transport in the Atmosphere-Vegetation-Soil Continuum " page 353 eq B.19}
\usage{
specific_hum2vpressure(q, p)
}
\arguments{
\item{q}{in [kg.kg^-1]}

\item{p}{in kPa}

\item{e}{in kPa}
}
\description{
vapour pressure in Pa after Moene and Van Dam Cambridge University Press 2014 "Transport in the Atmosphere-Vegetation-Soil Continuum " page 353 eq B.19
}
