% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{thermal_diff}
\alias{thermal_diff}
\title{thermal diffusivity in m^2/s. Table in page 50 of Moene and van Dam, "Transport in the Atmosphere-Vegetation-Soil Continuum "}
\usage{
thermal_diff()
}
\description{
thermal diffusivity in m^2/s. Table in page 50 of Moene and van Dam, "Transport in the Atmosphere-Vegetation-Soil Continuum "
}
