% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{vol_heat_capacity_soil}
\alias{vol_heat_capacity_soil}
\title{volumetric heat capacity of a soil. From the encyclopedia of soil science, page 306. Unit is J.m^-3.K^-1}
\usage{
vol_heat_capacity_soil(xw, xorg, xsolid)
}
\arguments{
\item{xw}{is the volumetric water content}

\item{xorg}{is the organic carbon content of the soil}

\item{xsolid}{is the mineral content of the soil}
}
\description{
volumetric heat capacity of a soil. From the encyclopedia of soil science, page 306. Unit is J.m^-3.K^-1
}
