% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{vpressure2specific_hum}
\alias{vpressure2specific_hum}
\title{specific humidity from vapour pressure in kg.kg-1 from formula above}
\usage{
vpressure2specific_hum(e, p)
}
\arguments{
\item{e}{in [Pa]}

\item{p}{in [Pa]}
}
\value{
q in [kg.kg^-1]
}
\description{
specific humidity from vapour pressure in kg.kg-1 from formula above
}
