% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowen.R
\name{wetbulb2vpressure}
\alias{wetbulb2vpressure}
\title{vapour pressure from wet bulb after Moene and Van Dam Cambridge University Press 2014 "Transport in the Atmosphere-Vegetation-Soil Continuum " page 354}
\usage{
wetbulb2vpressure(Twet, Tdry, p)
}
\arguments{
\item{Twet}{in [K]}

\item{Tdry}{in [K]}

\item{p}{in kPa}
}
\description{
vapour pressure from wet bulb after Moene and Van Dam Cambridge University Press 2014 "Transport in the Atmosphere-Vegetation-Soil Continuum " page 354
}
